+++ source/CHANGELOG.md
@@ -6,6 +6,9 @@ Version 2.4.6
 NEW:
 * Allow multiple files to be selected when importing files using the Import time series menu item 
 
+FIXED:
+* Analysis function Cumulative: now repects the input interpretation
+
 CHANGED:
 * Added progress bar to Calculator analysis function